+++ CHANGELOG.md
@@ -26,6 +26,12 @@ Fixed:
 
 - Resolved bug in regularization loss calculation for multi-sample batches
 
+## [XXXXX] - XXXX-XXX-XX
+
+Added:
+
+- `Diffusion map` as one niche trajectory construction method
+
 ## [1.0.7] - 2024-XXX-XX
 
 Added: